# Bench: optimise binary data generation