--- a/benches/read.rs
+++ b/benches/read.rs
@@ -51,7 +51,7 @@ fn benchmark(c: &mut Criterion) {
                 for i in 0_u64..100_000 {
                     writer
                         .send(match i {
-                            _ if i % 3 == 0 => Message::Binary(i.to_le_bytes().into()),
+                            _ if i % 3 == 0 => Message::binary(i.to_le_bytes().to_vec()),
                             _ => Message::Text(format!("{{\"id\":{i}}}")),
                         })
                         .unwrap();