--- a/benches/write.rs
+++ b/benches/write.rs
@@ -57,7 +57,7 @@ fn benchmark(c: &mut Criterion) {
         b.iter(|| {
             for i in 0_u64..100_000 {
                 let msg = match i {
-                    _ if i % 3 == 0 => Message::Binary(i.to_le_bytes().into()),
+                    _ if i % 3 == 0 => Message::binary(i.to_le_bytes().to_vec()),
                     _ => Message::Text(format!("{{\"id\":{i}}}")),
                 };
                 ws.write(msg).unwrap();